so the bug is fixed but, we may have to re do it.
So here’s the issue. How can we update the prescription table, without the doctor have to put in the first name and last name twice?

Because, if we don’t get a value from updating the prescription entity, we’ll likely get a parsing error.

So what to do? There could be something about joining tables perhaps, but I’m not sure.